Detect type mismatches between byte slice and code
If someone were to serialise a ciphertext of a particular N/W, and then deserialise into a different N/W ciphertext type, that would... not go well.
There are now tests to simulate at least basic forms of this sort of thing, and changes to catch these problems and return errors.