Rework KBKDF and nonce caching
On the KBKDF side, I was never happy with AES128 as the KBKDF, as it gets a bit shady if/when you need longer keys for something, and given how relatively infrequent key generation should be, the slight performance hit shouldn't be noticeable.  For the nonces, it was overkill to use the full KBKDF, seeded RNG will do the trick just fine.